removed list from utils, adding list tests
