+++ Program/src/filesystem_funcs.c
@@ -6,7 +6,7 @@
 #include <unistd.h>
 #include "utils.h"
 #include "string.h"
-
+#include "list.h"
 
 
 uint32_t get_file_size(char *source_file_name) {
