+++ Program/test/filesystem_tests.c
@@ -16,12 +16,12 @@ static receive_offset_tests(){
 
     receive_offset(file, 0, 5, 50);
     
-    file->missing_offsets->print(file->missing_offsets, nak_print, 0);
+    file->missing_offsets->iterate(file->missing_offsets, nak_print, 0);
     receive_offset(file, 0, 100, 1000);
 
-    file->missing_offsets->print(file->missing_offsets, nak_print, 0);
+    file->missing_offsets->iterate(file->missing_offsets, nak_print, 0);
     receive_offset(file, 0, 50, 100);
-    file->missing_offsets->print(file->missing_offsets, nak_print, 0);
+    file->missing_offsets->iterate(file->missing_offsets, nak_print, 0);
 
 }
