+++ Program/test/list_tests.c
@@ -1,0 +1,15 @@
+
+
+
+#include "list.h"
+#include "list_tests.h"
+
+int list_tests() {
+
+    List *list = linked_list();
+
+    
+
+
+
+}
