+++ Program/test/main.c
@@ -1,5 +1,4 @@
 
-#include "utils.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -16,12 +15,13 @@
 #include "filesystem_tests.h"
 #include "file_delivery_app.h"
 #include "protocol_handler_tests.h"
+#include "list_tests.h"
 
 Protocol_state *init_test() {
 
-Protocol_state *p_state = init_ftp(1);
-ssp_connectionless_server(p_state);
-ssp_cleanup_p_state(p_state);
+    Protocol_state *p_state = init_ftp(1);
+    ssp_connectionless_server(p_state);
+    ssp_cleanup_p_state(p_state);
 
 }
 
@@ -37,9 +37,10 @@ int main () {
     //ssp_cleanup_pdu_header(header);
     int error = 0;
 
-    error = request_tests();
-    error = protocol_handler_test();
-    error = packet_tests();
+    //error = request_tests();
+    //error = protocol_handler_test();
+    //error = packet_tests();
+    error = list_tests();
     //error = file_system_tests();
 
    return 0;
